Add files via upload
Predicting the price of used cars is an important and interesting problem. Most
individuals who buy new cars are also very worried about the resale value of
their cars after certain number of years when they will possibly sell it in the used
cars market.
Predicting the resale value of a car is not a simple task. We also have knowledge
that the value of used cars depends on a number of factors. The most important
ones are usually the age of the car, its make (and model), its mileage (the
number of kilometres it has run) and its horsepower.
The prices of new cars in the industry are fixed by the manufacturer with some
additional costs incurred by the Government in the form of taxes. So, customers
buying a new car can be assured of the money they invest to be worthy. But due
to the increased price of new cars and the incapability of customers to buy new
cars due to the lack of funds, used cars sales are on a global increase.
There is a need for a used car price prediction system to effectively determine
the worthiness of the car using a variety of features. Besides, different models
and systems may contribute on predicting power for a used car’s actual market
value. It is important to know their actual market value while both buying and
selling.